+++ addons/confirma/src/extensions/ConfirmStringExtensions.cs
@@ -5,6 +5,7 @@ namespace Confirma;
 
 public static class ConfirmStringExtensions
 {
+	#region ConfirmEmpty
 	public static void ConfirmEmpty(this string? actual, string? message = null)
 	{
 		if (string.IsNullOrEmpty(actual)) return;
@@ -24,7 +25,9 @@ public static class ConfirmStringExtensions
 			?? $"Expected non-empty string but found: {actual}"
 		);
 	}
+	#endregion
 
+	#region ConfirmContains
 	public static void ConfirmContains(
 		this string? actual,
 		string expected,
@@ -49,7 +52,9 @@ public static class ConfirmStringExtensions
 			?? $"Expected string to not contain: {expected} but found: {actual}"
 		);
 	}
+	#endregion
 
+	#region ConfirmStartsWith
 	public static void ConfirmStartsWith(this string? actual, string expected, string? message = null)
 	{
 		if (actual?.StartsWith(expected) == true) return;
@@ -69,7 +74,9 @@ public static class ConfirmStringExtensions
 			?? $"Expected string to not start with: {expected} but found: {actual}"
 		);
 	}
+	#endregion
 
+	#region ConfirmEndsWith
 	public static void ConfirmEndsWith(this string? actual, string expected, string? message = null)
 	{
 		if (actual?.EndsWith(expected) == true) return;
@@ -89,7 +96,9 @@ public static class ConfirmStringExtensions
 			?? $"Expected string to not end with: {expected} but found: {actual}"
 		);
 	}
+	#endregion
 
+	#region ConfirmHasLength
 	public static void ConfirmHasLength(this string? actual, int expected, string? message = null)
 	{
 		if (actual?.Length == expected) return;
@@ -99,4 +108,37 @@ public static class ConfirmStringExtensions
 			?? $"Expected string to have length: {expected} but found: {actual?.Length}"
 		);
 	}
+
+	public static void ConfirmNotHasLength(this string? actual, int expected, string? message = null)
+	{
+		if (actual?.Length != expected) return;
+
+		throw new ConfirmAssertException(
+			message
+			?? $"Expected string to not have length: {expected} but found: {actual?.Length}"
+		);
+	}
+	#endregion
+
+	#region ConfirmEqualsCaseInsensitive
+	public static void ConfirmEqualsCaseInsensitive(this string? actual, string expected, string? message = null)
+	{
+		if (string.Equals(actual, expected, StringComparison.OrdinalIgnoreCase)) return;
+
+		throw new ConfirmAssertException(
+			message
+			?? $"Expected string to equal: {expected} but found: {actual}"
+		);
+	}
+
+	public static void ConfirmNotEqualsCaseInsensitive(this string? actual, string expected, string? message = null)
+	{
+		if (!string.Equals(actual, expected, StringComparison.OrdinalIgnoreCase)) return;
+
+		throw new ConfirmAssertException(
+			message
+			?? $"Expected string to not equal: {expected} but found: {actual}"
+		);
+	}
+	#endregion
 }